+++ commander/sync_test.go
@@ -122,7 +122,7 @@ func (s *SyncTestSuite) TestSyncBatches_Transfer() {
 
 	latestBlockNumber, err := s.client.GetLatestBlockNumber()
 	s.NoError(err)
-	err = s.transactionExecutor.SyncBatches(0, uint64(*latestBlockNumber))
+	err = s.transactionExecutor.SyncBatches(0, *latestBlockNumber)
 	s.NoError(err)
 
 	txn, txStorage, err := s.storage.BeginTransaction(st.TxOptions{Postgres: true, Badger: true})
@@ -159,7 +159,7 @@ func (s *SyncTestSuite) TestSyncBatches_Transfer() {
 	s.NoError(err)
 	s.Len(batches, 1)
 
-	err = s.transactionExecutor.SyncBatches(0, uint64(*latestBlockNumber+2))
+	err = s.transactionExecutor.SyncBatches(0, *latestBlockNumber+2)
 	s.NoError(err)
 
 	state0, err := s.storage.GetStateLeaf(0)
@@ -213,7 +213,7 @@ func (s *SyncTestSuite) TestSyncBatches_Create2Transfer() {
 
 	latestBlockNumber, err := s.client.GetLatestBlockNumber()
 	s.NoError(err)
-	err = s.transactionExecutor.SyncBatches(0, uint64(*latestBlockNumber))
+	err = s.transactionExecutor.SyncBatches(0, *latestBlockNumber)
 	s.NoError(err)
 
 	state0, err := s.storage.GetStateLeaf(0)
