+++ eth/deployer/chain_connection.go
@@ -26,7 +26,7 @@ type ChainConnection interface {
 
 	GetChainID() models.Uint256
 
-	GetLatestBlockNumber() (*uint32, error)
+	GetLatestBlockNumber() (*uint64, error)
 
 	SubscribeNewHead(ch chan<- *types.Header) (ethereum.Subscription, error)
 }
